requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
Jinja2==3.1.2
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.3
python-dateutil==2.8.2
openai==1.9.0
